backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
PyJWT
email-validator